% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{add_colname_energy_units}
\alias{add_colname_energy_units}
\title{add_colname_energy_units()
returns unit extension in target columns according to the currently selected unit
of the app for power and energy related colnames. Should be called before making
nicely formatted columns with rename_fr_colnames()}
\usage{
add_colname_energy_units(data, unit)
}
\arguments{
\item{data}{the dataframe}

\item{unit}{the unit selected in the app}
}
\value{
dataframe with renamed columns
}
\description{
add_colname_energy_units()
returns unit extension in target columns according to the currently selected unit
of the app for power and energy related colnames. Should be called before making
nicely formatted columns with rename_fr_colnames()
}
