% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{add_colname_units}
\alias{add_colname_units}
\title{add_colnames_units
returns unit extension in target columns according to the currently selected unit
of the app for power and energy}
\usage{
add_colname_units(data, unit)
}
\arguments{
\item{data}{the dataframe}

\item{unit}{the unit selected in the app}
}
\value{
dataframe with renamed columns
}
\description{
add_colnames_units
returns unit extension in target columns according to the currently selected unit
of the app for power and energy
}
\examples{
"Consommation" will become "Consommation kWh" if unit = kWh,
and "Puissance" would become TJ/h if unit = TJ
}
