% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{batiment_danger}
\alias{batiment_danger}
\title{Buildings exposed to natural hazards}
\format{
A tibble with various rows and cols which
}
\usage{
batiment_danger
}
\description{
The dataset comes from DGE-UDN office and quantifies how many heated buildings
are exposed to mild/high natural hazards.
}
\keyword{datasets}
