% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_bar_plotly_dev}
\alias{create_bar_plotly_dev}
\title{create_bar_plotly()}
\usage{
create_bar_plotly_dev(
  data,
  n_communes,
  var_year,
  var_commune,
  unit,
  var_rank_2,
  var_values,
  color_palette,
  dodge = FALSE,
  free_y = FALSE,
  legend_title = NULL,
  web_width = 1500,
  web_height = 800,
  ...
)
}
\arguments{
\item{data}{the data to provide}
}
\value{
an interactive plotly object
}
\description{
Creates a plotly object from a facetted ggplot bar plot for use in renderPlotly
}
