% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_cons_table_dt}
\alias{create_cons_table_dt}
\title{create_cons_table_dt}
\usage{
create_cons_table_dt(data, unit)
}
\arguments{
\item{data}{Specific electricity consumption, DGE-DIREN data to transform to datatable.
Must follow specific data format which can be found in /data}
}
\value{
A DT table with export functionalities
}
\description{
create_cons_table_dt
}
