% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_doc_table_dt}
\alias{create_doc_table_dt}
\title{create_doc_table_dt
Creates minimalistic documentation table with download feature}
\usage{
create_doc_table_dt(data, doc_prefix)
}
\arguments{
\item{data}{}
}
\value{
A DT object
}
\description{
create_doc_table_dt
Creates minimalistic documentation table with download feature
}
