% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_generic_table_dt}
\alias{create_generic_table_dt}
\title{create_generic_table_dt
Creates datatable for generic datasets. Flexible enough to handle all datasets, but loses in
 customizability compared to other energy-specific datatable functions.}
\usage{
create_generic_table_dt(
  data,
  var_commune,
  var_year,
  var_cat = NULL,
  unit = unit,
  DT_dom = "Bfrtip"
)
}
\arguments{
\item{data}{the generic dataset}

\item{var_year}{the name of the year variable to sort from}

\item{DT_dom}{the datatable domain options to provide (e.g. 'Bfrtip')}
}
\value{
a DT object
}
\description{
create_generic_table_dt
Creates datatable for generic datasets. Flexible enough to handle all datasets, but loses in
 customizability compared to other energy-specific datatable functions.
}
