% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_geoportail_tag}
\alias{create_geoportail_tag}
\title{create_geoportail_tag
Creates a icon+link combination (tag) which redirects towards a specified geoportail link where geodata
can be viewed for Canton de Vaud}
\usage{
create_geoportail_tag(link)
}
\arguments{
\item{link}{the geoportail link that will be used inside the tag (https://geo.vd.ch/...)}
}
\value{
a span tag with icon and an html <a> tag with the redirect link
}
\description{
create_geoportail_tag
Creates a icon+link combination (tag) which redirects towards a specified geoportail link where geodata
can be viewed for Canton de Vaud
}
\examples{
create_geoportail_tag(link = "https://geo.vd.ch")
}
