% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_ng_cons_table_dt}
\alias{create_ng_cons_table_dt}
\title{create_ng_cons_table_dt
Creates datatable for ng_cons dataset}
\usage{
create_ng_cons_table_dt(
  data,
  var_commune,
  var_year,
  var_cat = NULL,
  energy_unit,
  DT_dom = "Bfrtip"
)
}
\arguments{
\item{data}{ng_cons dataset}

\item{var_year}{the name of the year variable to sort from}

\item{DT_dom}{the datatable domain options to provide (e.g. 'Bfrtip')}
}
\value{
a DT object
}
\description{
create_ng_cons_table_dt
Creates datatable for ng_cons dataset
}
