% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_plot_ggiraph}
\alias{create_plot_ggiraph}
\title{create_plot_ggiraph()}
\usage{
create_plot_ggiraph(
  data,
  n_communes,
  var_year,
  var_commune,
  unit,
  var_cat,
  var_values,
  geom,
  color_palette,
  dodge = FALSE,
  free_y = FALSE,
  legend_title = NULL,
  height_svg,
  width_svg
)
}
\arguments{
\item{data}{the data to provide}

\item{n_communes}{number of selected communes, used to control the width of facets}

\item{var_year}{the year variable}

\item{var_commune}{the commune variable}

\item{unit}{the unit to append in <var_values>}

\item{var_cat}{the optional categorical variable}

\item{var_values}{the variables containing the values}

\item{geom}{the type of geom for the plot : either 'col' or 'line'}

\item{color_palette}{a named vector of values-colors if <var_cat> is supplied, that should match <var_cat> items}

\item{dodge}{if geom = 'col' and <var_cat> is supplied : controls whether the cols are in a stacked or dodge position}

\item{free_y}{if <n_communes> is higher than 1, controls whether the y axis is independent for each facet or not}

\item{legend_title}{a string containing the legend title if <var_cat> is supplied}

\item{height_svg, width_svg}{dimensions of ggiraph output}
}
\value{
an interactive girafe object
}
\description{
Creates a girafe object from a facetted ggplot bar plot for use in renderGirafe
}
