% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_prod_table_dt}
\alias{create_prod_table_dt}
\title{create_table_dt}
\usage{
create_prod_table_dt(data, unit)
}
\arguments{
\item{data}{Specific electricity production, DGE-DIREN data to transform to datatable.
Must follow Pronovo's outputs and utils_helpers.R format.}
}
\value{
A DT table with export functionalities
}
\description{
create_table_dt
}
