% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_pv_density_plot}
\alias{create_pv_density_plot}
\title{create_pv_density_plot
DEVELOPMENT: Creates a density plot showing how PV installations installed power is distributed}
\usage{
create_pv_density_plot(data)
}
\arguments{
\item{data}{}
}
\value{
a ggplotly object
}
\description{
create_pv_density_plot
DEVELOPMENT: Creates a density plot showing how PV installations installed power is distributed
}
