% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_regener_table_dt}
\alias{create_regener_table_dt}
\title{create_regener_table_dt}
\usage{
create_regener_table_dt(data, unit)
}
\value{
a datatable object for regener datasets
}
\description{
create_regener_table_dt
}
