% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_subsidies_table_dt}
\alias{create_subsidies_table_dt}
\title{create_subsidies_table_dt()
Creates datatable for subsidies dataset}
\usage{
create_subsidies_table_dt(data, DT_dom = "Bfrtip")
}
\arguments{
\item{data}{the subsidies dataset}

\item{DT_dom}{datatable 'dom' Option, see datatable documentation. Likely Bfrtip or frtip}
}
\value{
a DT object
}
\description{
create_subsidies_table_dt()
Creates datatable for subsidies dataset
}
