% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_sunburst_plotly}
\alias{create_sunburst_plotly}
\title{create_sunburst_plotly}
\usage{
create_sunburst_plotly(
  data,
  year_var,
  year,
  values_tot,
  rank_1,
  rank_2,
  third_rank,
  rank_3_1,
  rank_3_2
)
}
\arguments{
\item{data}{the data to provide}
}
\value{
an interactive plot
}
\description{
Creates a sunburst, inspired from https://stackoverflow.com/questions/57395424/how-to-format-data-for-plotly-sunburst-diagram
}
