% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_treemap_plotly}
\alias{create_treemap_plotly}
\title{create_treemap_plotly}
\usage{
create_treemap_plotly(data)
}
\arguments{
\item{data}{the data to provide}
}
\value{
an interactive plot
}
\description{
Creates a treemap
}
