% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{elec_cons}
\alias{elec_cons}
\alias{elec_cons_doc}
\title{Electricity consumption by commune for canton de Vaud}
\format{
A tibble with various rows and cols which are not detailed yet

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\source{
DGE-DIREN(<https://www.vd.ch/>)
}
\usage{
elec_cons

elec_cons_doc
}
\description{
The dataset is an aggregation and simplification of all electricity delivery datapoints
which are yearly transmitted to DGE-DIREN by the distribution network managers.
Data is cleaned, affiliated to municipalities then aggregated and sent to the app.
}
\keyword{datasets}
