% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{format_numbers_heuristic}
\alias{format_numbers_heuristic}
\title{format_numbers_heuristic
a custom function to display large numbers (below 1000) without any decimal, and small numbers with one decimal
all values are displayed with thousand separator, no scientific notation and trailing zeroes are dropped}
\usage{
format_numbers_heuristic(number)
}
\arguments{
\item{number}{the number to format}
}
\value{
a formatted number as a string
}
\description{
format_numbers_heuristic
a custom function to display large numbers (below 1000) without any decimal, and small numbers with one decimal
all values are displayed with thousand separator, no scientific notation and trailing zeroes are dropped
}
