% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{generate_doc_accordion_panels}
\alias{generate_doc_accordion_panels}
\title{generate_doc_accordion_panels()
Create HTML `bslib::accordion_panel()` items for each dataset documented in ./data-doc. These are then used and
spliced as arguments inside a `bslib::accordion()` item. Each markdown h2 heading is used as a `title` argument for
`accordion_panel()` and each paragraph (i.e. not a h2 header) is used as content to fill the `accordion_panel()`.}
\usage{
generate_doc_accordion_panels(md_file)
}
\arguments{
\item{md_file}{the documentation filename stored in `./data-doc`}
}
\value{
a list of HTML accordion_panels to be spliced in a `bslib::accordion()`
}
\description{
generate_doc_accordion_panels()
Create HTML `bslib::accordion_panel()` items for each dataset documented in ./data-doc. These are then used and
spliced as arguments inside a `bslib::accordion()` item. Each markdown h2 heading is used as a `title` argument for
`accordion_panel()` and each paragraph (i.e. not a h2 header) is used as content to fill the `accordion_panel()`.
}
