% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{glossary}
\alias{glossary}
\title{Glossary}
\format{
A tibble with one column for the term, and one for the explanation.
}
\usage{
glossary
}
\description{
This dataset retrieves some specific terms which may not be obvious for lay people and provides explanations.
}
\keyword{datasets}
