% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{info_dev_message}
\alias{info_dev_message}
\title{info_dev_message()
shiny modal message to inform the user about the app, and offer the possibility to have a guided tour (introjs)}
\usage{
info_dev_message()
}
\value{
A a modal object when opening the app
}
\description{
info_dev_message()
shiny modal message to inform the user about the app, and offer the possibility to have a guided tour (introjs)
}
