% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{lump_alluvial_factors}
\alias{lump_alluvial_factors}
\title{lump_alluvial_factors()}
\usage{
lump_alluvial_factors(data, var_from, var_flow, var_to)
}
\arguments{
\item{var_to}{}
}
\description{
lump_alluvial_factors()
}
