% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{make_doc_table_dt}
\alias{make_doc_table_dt}
\title{make_doc_table_dt
Creates minimalistic documentation table with download features}
\usage{
make_doc_table_dt(data, doc_prefix)
}
\arguments{
\item{data}{the dataset containing variables and descriptions}
}
\value{
A DT object
}
\description{
make_doc_table_dt
Creates minimalistic documentation table with download features
}
