% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{make_slider_input_years}
\alias{make_slider_input_years}
\title{make_slider_input_years
creates a shiny sliderInput() widget based on a vector of two years}
\usage{
make_slider_input_years(id, years, ...)
}
\arguments{
\item{id}{the widget inputId, passed with a ns() to avoid namespace conflicts if inside a module}

\item{years}{a vector of two numeric years to be used as min, max, and default values}

\item{...}{<data-masking>}
}
\value{
a shiny sliderInput widget
}
\description{
make_slider_input_years
creates a shiny sliderInput() widget based on a vector of two years
}
\examples{
make_slider_input_years(id = "slider1", years = c(2010, 2025))
}
