% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{make_statbox_item}
\alias{make_statbox_item}
\title{make_statbox_item()
custom replacement for valueboxes (too rigid) to fill statboxes with icon, title, value and year info}
\usage{
make_statbox_item(iconBgClass, title, value, unit, year)
}
\value{
a HTML div container
}
\description{
make_statbox_item()
custom replacement for valueboxes (too rigid) to fill statboxes with icon, title, value and year info
}
\examples{
make_statbox_item(iconBgClass = NULL, title = "Rescues", value = 100, unit = "people", year = 2022)
}
