% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{make_table_dt}
\alias{make_table_dt}
\title{make_table_dt
a function that prepares data (add colnames, relocate, etc.) and makes a DT table
which also formats numbers nicely}
\usage{
make_table_dt(
  data,
  var_commune,
  var_year,
  var_values,
  var_cat,
  unit,
  na_string = "Non disponible",
  DT_dom = "frtip"
)
}
\arguments{
\item{data}{the dataframe or tibble to turn as a datatable}

\item{var_commune}{colname corresponding to the municipality (passed as a string)}

\item{var_year}{colname corresponding to the year (passed as a string)}

\item{var_values}{colname corresponding to the value(s) (passed as a string, or a vector of string)}

\item{var_cat}{colname corresponding to the categorical variable (passed as a string)}

\item{na_string}{a string specifying how should NAs be displayed, as it can change from one dataset to another}

\item{DT_dom}{the DT domain values to specify which DT extensions should be applied}

\item{string}{or variable containing a string or a list of strings, used as a prefix to `var_values` through `add_colname_unit`}
}
\value{
a dataTable object
}
\description{
make_table_dt
a function that prepares data (add colnames, relocate, etc.) and makes a DT table
which also formats numbers nicely
}
