% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{ng_cons}
\alias{ng_cons}
\alias{ng_cons_doc}
\title{Natural gas distribution by commune for canton de Vaud}
\format{
A tibble with various rows and cols which are not detailed yet

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\source{
DGE-DIREN(<https://www.vd.ch/>)
}
\usage{
ng_cons

ng_cons_doc
}
\description{
The dataset is an aggregation and simplification of all natural gas delivery datapoints
which are yearly transmitted to DGE-DIREN by the distribution network managers.
Data is cleaned, affiliated to municipalities then aggregated and sent to the app.
}
\keyword{datasets}
