% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{part_voit_elec}
\alias{part_voit_elec}
\title{Share of electric vehicles by commune for canton de Vaud}
\format{
A tibble with various rows and cols which are not detailed yet
}
\source{
DGMR (<https://www.vd.ch/>)
}
\usage{
part_voit_elec
}
\description{
The dataset comes from DGMR office and quantifies the share of (exclusively)
electric passenger cars among all the passenger cars registered in Canton de Vaud.
}
\keyword{datasets}
