% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{qualite_desserte}
\alias{qualite_desserte}
\title{Quality index for public transport services by commune for canton de Vaud}
\format{
A tibble with various rows and cols which
}
\usage{
qualite_desserte
}
\description{
The dataset comes from DGMR office and quantifies public transport services.
Two indices are calculated, by population and employment.
}
\keyword{datasets}
