% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{rename_columns_output}
\alias{rename_columns_output}
\title{rename_columns_output()
Uses the csv file in app/extdata/ to convert initial colnames to nicely formatted ones
the function matches only the start of the initial colname to find a match}
\usage{
rename_columns_output(data)
}
\arguments{
\item{data}{the dataset to rename with the default <colnames_replacement_display> object}
}
\value{
a renamed dataframe
}
\description{
rename_columns_output()
Uses the csv file in app/extdata/ to convert initial colnames to nicely formatted ones
the function matches only the start of the initial colname to find a match
}
