% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{rename_misc_colnames}
\alias{rename_misc_colnames}
\title{rename_misc_columns()
Specific function to rename specific columns that can't be fixed with `rename_fr_colnames()`
It relies on a named vector of specific columns to rename adequately.}
\usage{
rename_misc_colnames(data)
}
\value{
a renamed dataframe
}
\description{
rename_misc_columns()
Specific function to rename specific columns that can't be fixed with `rename_fr_colnames()`
It relies on a named vector of specific columns to rename adequately.
}
