% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{return_dynamic_size}
\alias{return_dynamic_size}
\title{return_dynamic_size
Returns a px value used for dynamic facet plots based on web display size and number of facets.
A facet row is typically well displayed at around 1/6 of the screen's height}
\usage{
return_dynamic_size(which, web_size, n_facets)
}
\arguments{
\item{which}{either 'width' or 'height'}

\item{web_size}{px size of either width or height, typically obtained with {shinybrowser}}
}
\value{
a numeric value corresponding to px
}
\description{
return_dynamic_size
Returns a px value used for dynamic facet plots based on web display size and number of facets.
A facet row is typically well displayed at around 1/6 of the screen's height
}
