% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{return_icons_subsidies}
\alias{return_icons_subsidies}
\title{return_icons_subsidies
returns one of two icons palette for subsidies data (either building or measure icons)}
\usage{
return_icons_subsidies(which)
}
\arguments{
\item{which}{either `building` to return the building palette,
or `measure` to return the measure palette}
}
\value{
a tibble with icon as html and a categorical variable to match icons
}
\description{
return_icons_subsidies
returns one of two icons palette for subsidies data (either building or measure icons)
}
