% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{return_palette_cons_elec}
\alias{return_palette_cons_elec}
\title{return_palette_cons_elec
Returns the color palette for sectors in the electricity consumption dataset}
\usage{
return_palette_cons_elec()
}
\value{
a vector with categorical data and hex color strings
}
\description{
return_palette_cons_elec
Returns the color palette for sectors in the electricity consumption dataset
}
