% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{return_palette_elec_cons}
\alias{return_palette_elec_cons}
\title{return_palette_elec_cons
Returns the color palette for sectors in the electricity consumption dataset}
\usage{
return_palette_elec_cons()
}
\value{
a vector with categorical data and hex color strings
}
\description{
return_palette_elec_cons
Returns the color palette for sectors in the electricity consumption dataset
}
