% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{return_palette_elec_prod}
\alias{return_palette_elec_prod}
\title{return_palette_elec_prod
Returns the color palette for categories in the electricity production dataset}
\usage{
return_palette_elec_prod()
}
\value{
a vector with categorical data and hex color strings
}
\description{
return_palette_elec_prod
Returns the color palette for categories in the electricity production dataset
}
