% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_communes}
\alias{sf_communes}
\title{Limits for Canton de Vaud - Communes}
\format{
A tibble with various rows and columns.
}
\source{
<https://viageo.ch/md/5be7ce8a-62b8-4031-8caa-5dfe7c0ef089>
}
\usage{
sf_communes
}
\description{
A simplified polygon layer for communes in Canton de Vaud, taken from viageo.ch.
}
\keyword{datasets}
