% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_layers}
\alias{sf_layers}
\alias{sf_canton}
\alias{sf_communes}
\alias{sf_districts}
\alias{sf_lacs}
\title{Limits for Canton de Vaud}
\format{
Each is a tibble with various rows and cols which are not detailed yet

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 300 rows and 8 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 6 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 3 columns.
}
\source{
<https://viageo.ch/md/5be7ce8a-62b8-4031-8caa-5dfe7c0ef089>
}
\usage{
sf_canton

sf_communes

sf_districts

sf_lacs
}
\description{
The different layers required to map the Canton de Vaud are stored in
`sf_canton`, `sf_communes`, `sf_districts` and `sf_lacs`. These are
simplified polygons (accuracy is not required here) taken from viageo.ch (see source)
and loaded as shapefiles with {sf} library
These layers must be updated when some merging between communes happens
}
\keyword{datasets}
