% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{subsidies_by_measure}
\alias{subsidies_by_measure}
\alias{subsidies_doc}
\title{Measure subsidies by commune for canton de Vaud}
\format{
A tibble with various rows and cols which are not detailed yet

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 4 columns.
}
\source{
<https://www.leprogrammebatiments.ch/fr>
}
\usage{
subsidies_by_measure

subsidies_doc
}
\description{
The dataset is a reflection of the main subsidies paid by Etat de Vaud for
renovation and changes of heat producers in buildings. The
program is called "Programme Bâtiments", and the primary data source is
IWF's platform used to process the subsidies. Data is acquired
through an authenticated API by DGE-DIREN, and data is pre-processed locally
before making it available to the application.
Energy-units should be expressed in kWh, and CO2 in tons.
}
\keyword{datasets}
