% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subsidies_yearly_state}
\alias{subsidies_yearly_state}
\title{Building subsidies by commune for Canton de Vaud}
\format{
A tibble with various rows and cols which are not detailed yet
}
\source{
<https://www.leprogrammebatiments.ch/fr>
}
\usage{
subsidies_yearly_state
}
\description{
The dataset is an agregation and simplification of all subsidies paid by
Etat de Vaud for renovation and changes of heat producers in buildings. The
programm is called "Programme Bâtiments", and the primary data source is
IWF's platform used to process the subsidies. Data is acquired
through a authentified API by DGE-DIREN, and data is pre-processed locally
before making it available to the application.
}
\examples{
subsidies_yearly_state
}
\keyword{datasets}
