% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{surface_canopee}
\alias{surface_canopee}
\title{Urban canopy surface by commune for canton de Vaud}
\format{
A tibble with various rows and cols which
}
\usage{
surface_canopee
}
\description{
The dataset comes from DGE-BIODIV office and quantifies how much of the
urban area is covered by more or less than 3m canopy height.
}
\keyword{datasets}
