% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{taux_motorisation}
\alias{taux_motorisation}
\title{Motorisation rate by commune for canton de Vaud}
\format{
A tibble with various rows and cols which are not detailed yet
}
\source{
DGMR (<https://www.vd.ch/>)
}
\usage{
taux_motorisation
}
\description{
The dataset comes from DGMR office and quantifies the number of motorised
vehicles per 1'000 inhabitants.
}
\keyword{datasets}
