% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{welcome_modal}
\alias{welcome_modal}
\title{welcome_modal()
shiny modal message to inform the user about the app, and offer the possibility to have a guided tour (introjs)}
\usage{
welcome_modal()
}
\value{
A a modal object when opening the app
}
\description{
welcome_modal()
shiny modal message to inform the user about the app, and offer the possibility to have a guided tour (introjs)
}
