% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{add_colname_unit}
\alias{add_colname_unit}
\title{add_colname_unit}
\usage{
add_colname_unit(data, colnames, unit)
}
\value{
dataframe with renamed columns (units added)
}
\description{
add_colname_unit
}
