% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_alluvial_chart}
\alias{create_alluvial_chart}
\title{create_alluvial_chart
creates a ggplot2 alluvial chart using the ggalluvial library and heat building
consumption data from an aggregated RegEner dataset}
\usage{
create_alluvial_chart(
  data,
  var_commune,
  var_flow,
  var_from,
  label_from,
  var_to,
  label_to
)
}
\value{
a ggplot2 object for regener alluvial visualisations
}
\description{
create_alluvial_chart
creates a ggplot2 alluvial chart using the ggalluvial library and heat building
consumption data from an aggregated RegEner dataset
}
