% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_plot_ggiraph}
\alias{create_plot_ggiraph}
\title{create_plot_ggiraph()}
\usage{
create_plot_ggiraph(
  data,
  n_communes,
  var_year,
  var_commune,
  unit,
  var_cat,
  var_values,
  geom,
  color_palette,
  dodge = FALSE,
  free_y = FALSE,
  legend_title = NULL,
  height_svg,
  width_svg,
  ...
)
}
\arguments{
\item{data}{the data to provide}
}
\value{
an interactive girafe object
}
\description{
Creates a girafe object from a facetted ggplot bar plot for use in renderGirafe
}
