% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_regener_table_dt}
\alias{create_regener_table_dt}
\title{create_regener_table_dt()
creates a DT table with custom formatting and html icons}
\usage{
create_regener_table_dt(data, unit, DT_dom = "Bfrtip")
}
\arguments{
\item{data}{Specific regener consumption table, DGE-DIREN data to transform to datatable}

\item{unit}{Unit currently selected inside the app}

\item{DT_dom}{datatable 'dom' Option, see datatable documentation. Likely Bfrtip or frtip}
}
\value{
a datatable object for regener datasets
}
\description{
create_regener_table_dt()
creates a DT table with custom formatting and html icons
}
