% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_rg_misc_table_dt}
\alias{create_rg_misc_table_dt}
\title{create_rg_misc_table_dt}
\usage{
create_rg_misc_table_dt(data)
}
\arguments{
\item{data}{the dataset containing variables and descriptions}
}
\value{
a DT object
}
\description{
create_rg_misc_table_dt
}
