% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_subsidies_table_dt}
\alias{create_subsidies_table_dt}
\title{create_subsidies_table_dt()
Creates datatable for subsidies dataset. Flexible enough to handle both datasets
`subsidies_by_building` and `subsidies_by_measure` hence its arguments for variables.}
\usage{
create_subsidies_table_dt(
  data,
  var_year,
  var_rank_2,
  icon_list,
  DT_dom = "Bfrtip"
)
}
\arguments{
\item{data}{the subsidies dataset}

\item{var_year}{the time variable}

\item{DT_dom}{datatable 'dom' Option, see datatable documentation. Likely Bfrtip or frtip}
}
\value{
a DT object
}
\description{
create_subsidies_table_dt()
Creates datatable for subsidies dataset. Flexible enough to handle both datasets
`subsidies_by_building` and `subsidies_by_measure` hence its arguments for variables.
}
