% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{create_sunburst_plotly}
\alias{create_sunburst_plotly}
\title{create_sunburst_plotly}
\usage{
create_sunburst_plotly(
  data_sunburst,
  unit,
  var_year,
  var_values,
  var_commune,
  var_rank_2,
  third_rank,
  var_rank_3_1,
  var_rank_3_2
)
}
\arguments{
\item{data}{the data to provide}
}
\value{
an interactive plot
}
\description{
Creates a sunburst, inspired from https://stackoverflow.com/questions/57395424/how-to-format-data-for-plotly-sunburst-diagram
}
