% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elec_prod}
\alias{elec_prod}
\alias{elec_prod_doc}
\title{Electricity production by commune for Canton de Vaud}
\format{
A tibble with various rows and cols which are not detailed yet

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 4 columns.
}
\source{
<https://www.pronovo.ch/fr>
}
\usage{
elec_prod

elec_prod_doc
}
\description{
The dataset is an agregation and simplification of all electricity production
installations recorded in canton de Vaud by PRONOVO AG. The data is split by
category of installation (i.e. photovoltaïcs, wind turbines, etc.). The data
results in a fairly complex methodology where DGE-DIREN provides some estimates
for some type of installation which production is not recorded, or for self-consumption.
The data is accompanied with a separate dictionnary dataset `elec_prod_doc`.
}
\keyword{datasets}
