% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{info_dev_message}
\alias{info_dev_message}
\title{info_dev_message()
shinyalert popup message from the dev to explain what is the app made for}
\usage{
info_dev_message()
}
\value{
A shinyalert object when opening the app
}
\description{
info_dev_message()
shinyalert popup message from the dev to explain what is the app made for
}
